requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
aiosmtplib==2.0.2
alembic==1.12.0
annotated-types==0.6.0
anyio==3.7.1
attrs==23.2.0
azure-core==1.30.1
azure-storage-blob==12.19.1
bcrypt==4.1.2
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
dataclasses-json==0.6.4
distro==1.9.0
dnspython==2.6.1
ecdsa==0.18.0
email_validator==2.1.1
fastapi==0.103.2
fastapi-mail==1.4.1
frozenlist==1.4.1
h11==0.14.0
httpcore==0.17.3
httptools==0.6.1
httpx==0.24.1
idna==3.6
iniconfig==2.0.0
isodate==0.6.1
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.12
langchain-community==0.0.28
langchain-core==0.1.32
langchain-openai==0.0.8
langchain-pinecone==0.0.3
langchain-text-splitters==0.0.1
langsmith==0.1.30
Mako==1.3.2
MarkupSafe==2.1.5
marshmallow==3.21.1
multidict==6.0.5
mypy-extensions==1.0.0
mysql-connector-python==8.3.0
numpy==1.26.4
openai==1.14.2
orjson==3.9.15
packaging==23.2
passlib==1.7.4
pdfminer.six==20231228
pdfplumber==0.11.0
pillow==10.2.0
pinecone-client==3.1.0
pluggy==1.4.0
pyasn1==0.5.1
pycparser==2.21
pydantic==2.6.3
pydantic-settings==2.0.3
pydantic_core==2.16.3
PyMySQL==1.1.0
pypdfium2==4.28.0
pytest==8.1.1
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-toolbelt==1.0.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.28
starlette==0.27.0
tenacity==8.2.3
tiktoken==0.6.0
tqdm==4.66.2
typing-inspect==0.9.0
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.23.2
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
yarl==1.9.4
zep-python==1.5.0
